osv: xen: move common code outside of arch/x64
Split arch/x64/xen.hh into common and x64-specific parts,
fixing inclusions along the way.
Move xen_intr.cc to core, as it's mostly platform-agnostic.
AARCH64 will be able to reuse this code and headers.
