+++ arch/aarch64/early-console.cc
@@ -6,9 +6,9 @@
  */
 
 #include <osv/prio.hh>
+#include <osv/xen.hh>
 
 #include "early-console.hh"
-#include "xen.hh"
 
 namespace console {
